[GprReactor] Write and read-in all nuclides, remove case-specific variables
Removed variables include sets of relevant nuclides and of permitted
nuclides, as well as specific variables such as temperature, which
were used previously in a specific case study.

In the future, I might add a user-defined variable that allows to
specify any inputs that could be relevant to the GPRs (e.g., some
temperatures) but not to Cyclus. With the changes implemented in this
commit, only variables actually relevant to the Cyclus-side of
GprReactor (such as the cycle time) are passed to the Gpr-JSON input
file, as well.