[GprReactor] Change to NVD's own GPR model, increase flexibility.
The GP models used are from https://github.com/jung-benjamin/gp-surfer.

This commit increases flexibility in the sense that the user can
(must) define which parameters are used to perform the GP
calculations -- before, this archetype and the GP calculations
were very restrictive. This is probably also a first step in
addressing issue #14.